%    Copyright (c) 2012-2018 Russell V. Lenth                                %

\name{auto.noise}
\alias{auto.noise}
\alias{feedlot}
\alias{fiber}
\alias{MOats}
\alias{nutrition}
\alias{oranges}

\docType{data}
\title{Data sets}
\description{The datasets \sQuote{auto.noise}, \sQuote{feedlot}, \sQuote{fiber},
\sQuote{MOats}, \sQuote{nutrition}, and \sQuote{oranges} are provided in case a user customarily loads the data from \pkg{lsmeans}. But the same datasets are provided in the \pkg{emmeans} package, and they are documented there.}

\usage{auto.noise}

\author{Russell V. Lenth}
\keyword{datasets}
