% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalizeFirst.R
\name{capitalizeFirst}
\alias{capitalizeFirst}
\title{capitalizeFirst function}
\usage{
capitalizeFirst(name)
}
\arguments{
\item{name}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
