% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBlupH2.R
\name{fitBlupH2}
\alias{fitBlupH2}
\title{fitlsmeans function}
\usage{
fitBlupH2(form, effect, addMu = FALSE, ...)
}
\arguments{
\item{form}{object of class formula, compliant with lmer().}

\item{addMu}{logical. Should the mean be added back into the BLUP? Default is FALSE}

\item{data}{data.frame.  Name of effect for which lsmeans should be calculated}
}
\value{
list. Returns list of BLUP, mean, and h2, i.e. heritability
}
\description{
function to fit a linear mixed model and calculate BLUPs and heritability.
}
\details{
extracts BlUPs and calculates heritability 
This function is a wrapper for getBlupH2, and was made to calculate iid BLUPs and heritability
}
\examples{
fitBlupH2(mpg ~ 1|cyl, effect = "cyl", data = mtcars)
fitBlupH2(Sepal.Length ~ 1|Species, "Species", data = iris)
}
