\name{gDGC-package}
\alias{gDGC-package}
\alias{gDGC}
\docType{package}
\title{gDGC}
\description{The gDGC package provides routines to perform the univariate and multivariate versions of the DGC 
             multiple-means-comparison method (Di Rienzo, et al 2002; Valdano and Di Rienzo, 2007).
}
\details{
\tabular{ll}{
Package: \tab gDGC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-03\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}


The package contains the following main functions: gDGC and MDGCQ.

}
\author{
	Julio A. Di Rienzo  <email: dirienzo@agro.unc.edu.ar> and Maria del Carmen Romero <email: mariadelc.romero@gmail.com>
	
	Maintainer: 
	Julio A. Di Rienzo  <email: dirienzo@agro.unc.edu.ar> and Maria del Carmen Romero <email: mariadelc.romero@gmail.com>
 }
\references{

}
\keyword{ package }
\seealso{
	\code{\link{gDGC}}, 
	\code{\link{MDGCQ}},


}
\examples{

}
