% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlupH2.R
\name{getBlupH2}
\alias{getBlupH2}
\title{getBlupH2 function}
\usage{
getBlupH2(lmerFit, effect, addMu = FALSE)
}
\arguments{
\item{lmerFit}{[value]}

\item{effect}{[value]}

\item{addMu}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
