% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBLUtab.R
\name{makeBLUtab}
\alias{makeBLUtab}
\title{makeBLUtab function}
\usage{
makeBLUtab(BLU, sortby = NULL, addInfo = NULL)
}
\arguments{
\item{BLU}{[value]}

\item{sortby}{[value]. Default is NULL}

\item{addInfo}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
