% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBLUs.R
\name{mergeBLUs}
\alias{mergeBLUs}
\title{mergeBLUs function}
\usage{
mergeBLUs(BLUlist, sortHiLo = NULL, addInfo)
}
\arguments{
\item{BLUlist}{[value]}

\item{sortHiLo}{[value]. Default is NULL}

\item{addInfo}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
