\docType{class}
\name{ref.grid-class}
\alias{ref.grid-class}
\alias{lsmobj-class}

\title{The \code{ref.grid} and \code{lsmobj} classes}

\description{
The codebase for \pkg{lsmeans} is now mostly in \pkg{emmeans}. These two classes
are simple extensions of the \code{emmGrid} class defined in \pkg{emmeans},
and are provided as support for objects created in older versions of \pkg{lsmeans}.
For details, see \code{\link{emmGrid-class}}.
}

\author{Russell V. Lenth}
