% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialUnion.R
\name{trialUnion}
\alias{trialUnion}
\title{trialUnion function}
\usage{
trialUnion(trialList, resetRowNames = TRUE)
}
\arguments{
\item{trialList}{[value]}

\item{resetRowNames}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
