% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlupH2.R
\name{getBlupH2}
\alias{getBlupH2}
\title{getBlupH2 function}
\usage{
getBlupH2(lmerFit, effect, addMu = FALSE)
}
\arguments{
\item{lmerFit}{lmer model object.}

\item{effect}{factor to extract BLUPs and calculate heritability}

\item{addMu}{logical. Should the mean be added back into the BLUP? Default is FALSE}
}
\value{
list. Returns list of BLUP, mean, and h2, i.e. heritability
}
\description{
function to extract BLUPs and calculate heritability.
}
\details{
extracts BlUPs and calculates heritability
}
\examples{
none
}
