\name{lsmeans-package}
\alias{lsmeans-package}
\docType{package}
\title{
Least-squares means
}
\description{
This package provides the function \code{\link{lsmeans}} for outputting so-called least-squares means for factor combinations in an \code{lm}, \code{lme}, or \code{lmer} object. It can also compute contrasts or linear combinations of these least-squares means, and several standard contrast families are provided.
Examples and discussion are available via \code{vignette("lsmeans", package="lsmeans")}.
}
\details{
\tabular{ll}{
Package: \tab lsmeans\cr
Type: \tab Package\cr
Version: \tab See DESCRIPTION\cr
Date: \tab See DESCRIPTION\cr
License: \tab GPL-2\cr
}
See the documentation for \code{\link{lsmeans}}.
}
\author{
Russell V. Lenth

Maintainer: Russ Lenth <russell-lenth@uiowa.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\keyword{ models }
\keyword{ regression }
\keyword{ htest }
\seealso{
\code{\link[nlme:lme]{nlme}} ~~
\code{\link[lme4:lmer]{lme4}} ~~
}
