% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{lsmeans}
\alias{lsmeans}
\title{lsmeans function}
\usage{
lsmeans(lmfit, effect, meanType = "intercept", alpha = 0.05)
}
\arguments{
\item{lmfit}{[value]}

\item{effect}{[value]}

\item{meanType}{[value]. Default is "intercept"}

\item{alpha}{[value]. Default is 0.05}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
